Updating to latest Nyquist

+++ sf-cvs/trunk/lib-src/libnyquist/tran/pluck.c
@@ -1,5 +1,3 @@
-#include <math.h>
-
 #include "stdio.h"
 #ifndef mips
 #include "stdlib.h"
@@ -107,7 +105,7 @@ void pluck_initialize(sample_type *shiftreg, sample_type *array,
            values to +1 or -1. The following does the same
            thing with much less code:
          */
-        *array = (krand() & 2) - 1;
+        *array = (float) ((krand() & 2) - 1);
         suma += *array; /* compute sum for the average */
     }
     avea = suma / len;
@@ -142,33 +140,33 @@ void pluck__fetch(register pluck_susp_type susp, snd_list_type snd_list)
     snd_list->block = out;
 
     while (cnt < max_sample_block_len) { /* outer loop */
-        /* first compute how many samples to generate in inner loop: */
-        /* don't overflow the output sample block: */
-        togo = max_sample_block_len - cnt;
+	/* first compute how many samples to generate in inner loop: */
+	/* don't overflow the output sample block: */
+	togo = max_sample_block_len - cnt;
 
-        /* don't run past terminate time */
-        if (susp->terminate_cnt != UNKNOWN &&
-            susp->terminate_cnt <= susp->susp.current + cnt + togo) {
-            togo = susp->terminate_cnt - (susp->susp.current + cnt);
-            if (togo == 0) break;
-        }
+	/* don't run past terminate time */
+	if (susp->terminate_cnt != UNKNOWN &&
+	    susp->terminate_cnt <= susp->susp.current + cnt + togo) {
+	    togo = susp->terminate_cnt - (susp->susp.current + cnt);
+	    if (togo == 0) break;
+	}
 
-        n = togo;
-        stretch_reg = susp->stretch;
-        cons_reg = susp->cons;
-        loss_reg = susp->loss;
-        x2_reg = susp->x2;
-        x3_reg = susp->x3;
-        i1_reg = susp->i1;
-        i2_reg = susp->i2;
-        i3_reg = susp->i3;
-        i4_reg = susp->i4;
-        endptr_reg = susp->endptr;
-        out_ptr_reg = out_ptr;
-        if (n) do { /* the inner sample computation loop */
-            sample_type sum =
-                (*i1_reg++ * x2_reg) + (*i2_reg++ * x3_reg) + 
-                (*i3_reg++ * stretch_reg) - (*i4_reg++ * cons_reg);
+	n = togo;
+	stretch_reg = susp->stretch;
+	cons_reg = susp->cons;
+	loss_reg = susp->loss;
+	x2_reg = susp->x2;
+	x3_reg = susp->x3;
+	i1_reg = susp->i1;
+	i2_reg = susp->i2;
+	i3_reg = susp->i3;
+	i4_reg = susp->i4;
+	endptr_reg = susp->endptr;
+	out_ptr_reg = out_ptr;
+	if (n) do { /* the inner sample computation loop */
+            sample_type sum = (float)
+                ((*i1_reg++ * x2_reg) + (*i2_reg++ * x3_reg) + 
+                 (*i3_reg++ * stretch_reg) - (*i4_reg++ * cons_reg));
             /* wrap pointers around shift register if necessary */
             if (i1_reg == endptr_reg) i1_reg = susp->shiftreg;
             if (i2_reg == endptr_reg) i2_reg = susp->shiftreg;
@@ -176,27 +174,27 @@ void pluck__fetch(register pluck_susp_type susp, snd_list_type snd_list)
             if (i4_reg == endptr_reg) i4_reg = susp->shiftreg;
 
             /* store new value in shift register */
-            *i4_reg = sum * loss_reg;
+            *i4_reg = (float) (sum * loss_reg);
 
             /* deliver sample */
             *out_ptr_reg++ = sum;
 ;
-        } while (--n); /* inner loop */
+	} while (--n); /* inner loop */
 
-        susp->i1 = i1_reg;
-        susp->i2 = i2_reg;
-        susp->i3 = i3_reg;
-        susp->i4 = i4_reg;
-        out_ptr += togo;
-        cnt += togo;
+	susp->i1 = i1_reg;
+	susp->i2 = i2_reg;
+	susp->i3 = i3_reg;
+	susp->i4 = i4_reg;
+	out_ptr += togo;
+	cnt += togo;
     } /* outer loop */
 
     /* test for termination */
     if (togo == 0 && cnt == 0) {
-        snd_list_terminate(snd_list);
+	snd_list_terminate(snd_list);
     } else {
-        snd_list->block_len = cnt;
-        susp->susp.current += cnt;
+	snd_list->block_len = cnt;
+	susp->susp.current += cnt;
     }
 } /* pluck__fetch */
